+++ lib/src/working_copy.rs
@@ -31,6 +31,8 @@ use crate::commit::Commit;
 use crate::conflicts::ConflictMarkerStyle;
 use crate::dag_walk;
 use crate::fsmonitor::FsmonitorSettings;
+use crate::gitattributes::GitAttributesError;
+use crate::gitattributes::GitAttributesFile;
 use crate::gitignore::GitIgnoreError;
 use crate::gitignore::GitIgnoreFile;
 use crate::matchers::EverythingMatcher;
@@ -184,6 +186,9 @@ pub enum SnapshotError {
     /// Checking path with ignore patterns failed.
     #[error(transparent)]
     GitIgnoreError(#[from] GitIgnoreError),
+    /// Checking path with gitattributes patterns failed.
+    #[error(transparent)]
+    GitAttributesError(#[from] GitAttributesError),
     /// Some other error happened while snapshotting the working copy.
     #[error("{message}")]
     Other {
@@ -205,6 +210,8 @@ pub struct SnapshotOptions<'a> {
     // because the TreeState may be long-lived if the library is used in a
     // long-lived process.
     pub base_ignores: Arc<GitIgnoreFile>,
+    /// Used for ignoring LFS files - if the setting isn't enabled, this is None
+    pub base_attributes: Arc<GitAttributesFile>,
     /// The fsmonitor (e.g. Watchman) to use, if any.
     // TODO: Should we make this a field on `LocalWorkingCopy` instead since it's quite specific to
     // that implementation?
@@ -229,6 +236,7 @@ impl SnapshotOptions<'_> {
     pub fn empty_for_test() -> Self {
         SnapshotOptions {
             base_ignores: GitIgnoreFile::empty(),
+            base_attributes: Arc::new(GitAttributesFile::default()),
             fsmonitor_settings: FsmonitorSettings::None,
             progress: None,
             start_tracking_matcher: &EverythingMatcher,